# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CMS",
        "description": "Todas las rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/v1/deportes": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Obtener una lista de deportes",
                "operationId": "68a9d1d3d21c267062e073ab25cbc2f7",
                "responses": {
                    "200": {
                        "description": "Lista de deportes obtenida con éxito"
                    },
                    "404": {
                        "description": "No se han encontrado deportes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Registrar un nuevo deporte",
                "operationId": "0b8b57ebd07a881d07e87cda92e1f99b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Baloncesto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte registrado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "500": {
                        "description": "Deporte existente, pruebe otro deporte"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/deportes/{id}": {
            "get": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Obtener detalles de un deporte específico",
                "operationId": "e2bc535a8d2149036be8a6a54a17ca29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del deporte obtenidos con éxito"
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Actualizar los datos de un deporte",
                "operationId": "08cc7c5bccc1729b8c4612eddebad682",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Baloncesto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deporte actualizado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Deportes"
                ],
                "summary": "Eliminar un deporte",
                "operationId": "e0bc6c8e247aeb2cc0cef2cb686aa0e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deporte eliminado con éxito"
                    },
                    "403": {
                        "description": "No tiene permisos para realizar esta acción"
                    },
                    "404": {
                        "description": "Deporte no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token de acceso para autenticar las peticiones",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}